ai-agent/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
google-genai
python-dotenv
